Getting rid of "long" data type. As far as I see it, none of these changes really fix anything.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7002 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/DarkmodAASHidingSpotFinder.cpp
@@ -1050,9 +1050,9 @@ void CDarkmodAASHidingSpotFinder::debugAppendHidingSpotsToDraw
 {
 	idList<darkModHidingSpot> hidingSpotsToAppend;
 
-	unsigned long numSpots = inout_hidingSpots.getNumSpots();
+	unsigned int numSpots = inout_hidingSpots.getNumSpots();
 
-	for (unsigned long spotIndex = 0; spotIndex < numSpots; spotIndex ++)
+	for (unsigned int spotIndex = 0; spotIndex < numSpots; spotIndex ++)
 	{
 		darkModHidingSpot* p_spot = inout_hidingSpots.getNthSpot (spotIndex);
